requirements.txt
arrow==1.2.3
asgiref==3.5.1
async-timeout==4.0.2
click==8.1.3
colorama==0.4.5
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.16
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-import-export==3.0.1
django-redis==5.2.0
django-rq==2.5.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
rq==1.11.1
git+https://github.com/michaelbrooks/rq-win.git#egg=rq-win
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.4.2
tablib==3.2.1
times==0.7
tzdata==2022.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0